FastAPI_backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
dynaconf==3.2.6
pydantic[email]
python-multipart
PyJWT